fix: getGrids and pr changes
